Listen for confirmation
In the updated code, I have added the listenForConfirmation function, which retrieves the transaction receipt using contractKit.web3.eth.getTransactionReceipt and checks for the transaction status. If the receipt is not available, it will continue to poll every 3 seconds until the receipt is available.

Make sure to replace '0xrecipientAddress' with the actual recipient address where you want to send the transaction.